Added Delete and Update functionalities, Error handling pending

+++ models/Note.js
@@ -26,19 +26,19 @@ noteSchema.statics.getNotes = async function (user_id) {
     const notes = await this.find({ user_id: user_id });
     return notes;
 }
-noteSchema.statics.getNote = async function (note_id) {
+noteSchema.statics.getNote = async function (note_id,user_id) {
     try {
 
-        const note = await this.findOne({ _id: note_id });
+        const note = await this.findOne({ _id: note_id,user_id:user_id });
         return note;
     }
     catch (err) {
         console.log(err);
     }
 }
-noteSchema.statics.deleteNote = async function (note_id) {
+noteSchema.statics.deleteNote = async function (note_id,user_id) {
     try {
-        await this.deleteOne({ _id: note_id }); 
+        await this.deleteOne({ _id: note_id,user_id:user_id }); 
         return true;
     }
     catch (err) {
@@ -46,6 +46,16 @@ noteSchema.statics.deleteNote = async function (note_id) {
     }
 
 }
+noteSchema.statics.update = async function (note_id,newText,user_id) {
+    try {
+        const note = await this.findOneAndUpdate({  "_id": note_id,"user_id":user_id  },{"text" : newText },{useFindAndModify:false}); 
+        return note;
+    }
+    catch (err) {
+        console.log(err.message);
+    }
+}
+
 
 
 const Note = mongoose.model('note', noteSchema);
